+++ config/localization.php
@@ -3,6 +3,12 @@
 return [
 
     /**
+     * Applications default locale need to be set because the config('app.locale')
+     * gives back the current locale and not the value from config
+     */
+    'default_locale' => 'de',
+
+    /**
      * Application locales determines all locals that exists in the application
      */
     'locales' => [
@@ -17,6 +23,7 @@ return [
 
     /**
      * Detect user locale via http header
+     * When no locale is stored in session user gets redirected
      */
     'detect_locale' => true,